101— Lifting the state up
State moved from child to parent so to say.
Parent (App) is providing dynamic data to input component so they can update their values. And also for Results component to display the data dynamically.
Everything is clear, he is doing slightly different but nothing is strange.
I am lovin' it.